harden server and CLI: host check, headers, port, -f, exit codes, tty hints, confirm passphrase

+++ c/find.h
@@ -2,8 +2,8 @@
  *
  * ais_find prints every live record whose value contains NEEDLE (a plain
  * substring, case-sensitive) as "id|value", one line per matching value.
- * Tombstoned records are suppressed. Streaming, bounded memory. Returns 0, or
- * -1 on error.
+ * Tombstoned records are suppressed. Streaming, bounded memory. Returns how
+ * many lines matched, or -1 on error.
  */
 #ifndef AIS_FIND_H
 #define AIS_FIND_H
@@ -11,6 +11,6 @@
 #include <stdio.h>
 #include "ais.h"
 
-int ais_find(ais *a, const char *needle, FILE *out);
+long ais_find(ais *a, const char *needle, FILE *out);
 
 #endif /* AIS_FIND_H */
